Working branch

+++ src/main/java/equation/app/repository/AnswerRepository.java
@@ -1,6 +1,7 @@
 package equation.app.repository;
 
 import equation.app.model.Answer;
+import java.math.BigDecimal;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -10,5 +11,5 @@ import org.springframework.stereotype.Repository;
 public interface AnswerRepository extends JpaRepository<Answer, Long> {
 
     @Query("from Answer a left join fetch a.equations where a.value = ?1")
-    Optional<Answer> findByValue(Double value);
+    Optional<Answer> findByValue(BigDecimal value);
 }
